CordovaFiniteObservable moved to photo-library. Overloads are marked as private.
